+++ packages/plugin-packages/mashroom-session-provider-mongodb/src/mongodb.ts
@@ -7,7 +7,7 @@ import type {MashroomLogger} from '@mashroom/mashroom/type-definitions';
 let _connectionUri: string | null = null;
 let _connectionOptions: MongoClientOptions | null = null;
 let _client: MongoClient | null = null;
-let _lastHeartbeatSucceeded = false;
+let _topologyDescription: TopologyDescription | undefined;
 
 export const close = async (): Promise<void> => {
     if (_client) {
@@ -26,24 +26,17 @@ export const setConnectionUriAndOptions = async (connectionUri: string, options:
     _connectionOptions = options;
 };
 
-const getTopology = (): TopologyDescription | undefined => {
-    // @ts-ignore Accessing the private property topology
-    return _client?.topology?.description;
-};
-
 export const getAvailableNodes = () => {
-    const topology = getTopology();
-    if (!topology) {
-        // Fallback
-        return _lastHeartbeatSucceeded ? 1 : 0;
+    if (!_topologyDescription) {
+        return 0;
     }
-    let servers = 0;
-    topology.servers.forEach((server) => {
-        if (!server.error) {
-            servers ++;
+    let availableNodes = 0;
+    _topologyDescription.servers.forEach((server) => {
+        if (server.type !== 'Unknown' && !server.error) {
+            availableNodes ++;
         }
     });
-    return servers;
+    return availableNodes;
 };
 
 export const isConnected = () => {
@@ -70,19 +63,7 @@ export default async (logger: MashroomLogger): Promise<MongoClient> => {
 
     _client.on('topologyDescriptionChanged', (event) => {
         logger.info('MongoDB: New topology description:', event.newDescription);
-    });
-
-    _client.on('serverDescriptionChanged', (event) => {
-        logger.info('MongoDB: New server description:', event.newDescription);
-    });
-
-    _client.on('serverHeartbeatSucceeded', (event) => {
-        _lastHeartbeatSucceeded = true;
-        logger.debug('MongoDB: Heartbeat succeeded:', event);
-    });
-    _client.on('serverHeartbeatFailed', (event) => {
-        _lastHeartbeatSucceeded = false;
-        logger.error('MongoDB: Heartbeat failed:', event);
+        _topologyDescription = event.newDescription;
     });
 
     await _client.connect();